+++ x/leverage/keeper/grpc_query.go
@@ -238,27 +238,38 @@ func (q Querier) AccountSummary(
 	borrowed := q.Keeper.GetBorrowerBorrows(ctx, addr)
 
 	// the following price calculations use the most recent prices if spot prices are missing
-	lastSuppliedValue, err := q.Keeper.VisibleTokenValue(ctx, supplied, types.PriceModeLast)
+	lastSuppliedValue, err := q.Keeper.VisibleTokenValue(ctx, supplied, types.PriceModeQuery)
 	if err != nil {
 		return nil, err
 	}
-	lastBorrowedValue, err := q.Keeper.VisibleTokenValue(ctx, borrowed, types.PriceModeLast)
+	lastBorrowedValue, err := q.Keeper.VisibleTokenValue(ctx, borrowed, types.PriceModeQuery)
 	if err != nil {
 		return nil, err
 	}
-	lastCollateralValue, err := q.Keeper.VisibleCollateralValue(ctx, collateral, types.PriceModeLast)
+	lastCollateralValue, err := q.Keeper.VisibleCollateralValue(ctx, collateral, types.PriceModeQuery)
 	if err != nil {
 		return nil, err
 	}
 
-	// this supplied value uses leverage-like prices: the lower of spot or historic price for supplied tokens
-	suppliedValue, err := q.Keeper.VisibleTokenValue(ctx, supplied, types.PriceModeLow)
+	// these use leverage-like prices: the lower of spot or historic price for supplied tokens and higher for borrowed.
+	// unlike transactions, this query will use expired prices instead of skipping them.
+	suppliedValue, err := q.Keeper.VisibleTokenValue(ctx, supplied, types.PriceModeQueryLow)
+	if err != nil {
+		return nil, err
+	}
+	collateralValue, err := q.Keeper.VisibleCollateralValue(ctx, collateral, types.PriceModeQueryLow)
+	if err != nil {
+		return nil, err
+	}
+	borrowedValue, err := q.Keeper.VisibleTokenValue(ctx, borrowed, types.PriceModeQueryHigh)
 	if err != nil {
 		return nil, err
 	}
 
 	resp := &types.QueryAccountSummaryResponse{
 		SuppliedValue:       suppliedValue,
+		CollateralValue:     collateralValue,
+		BorrowedValue:       borrowedValue,
 		SpotSuppliedValue:   lastSuppliedValue,
 		SpotCollateralValue: lastCollateralValue,
 		SpotBorrowedValue:   lastBorrowedValue,
@@ -274,9 +285,9 @@ func (q Querier) AccountSummary(
 		return nil, err
 	}
 	if err == nil {
-		resp.BorrowedValue = ap.BorrowedValue()
-		resp.CollateralValue = ap.CollateralValue()
-		resp.BorrowLimit = ap.Limit()
+		// on missing borrow price, borrow limit is nil
+		borrowLimit := ap.Limit()
+		resp.BorrowLimit = &borrowLimit
 	}
 
 	// liquidation threshold shown here as it is used in leverage logic: using spot prices.
@@ -287,7 +298,7 @@ func (q Querier) AccountSummary(
 		return nil, err
 	}
 	if err == nil {
-		// on an error here, simply skip setting the response field
+		// on missing collateral price, liquidation threshold is nil
 		liquidationThreshold := ap.Limit()
 		resp.LiquidationThreshold = &liquidationThreshold
 	}