+++ x/leverage/keeper/oracle.go
@@ -29,10 +29,10 @@ func (k Keeper) TokenPrice(ctx sdk.Context, baseDenom string, mode types.PriceMo
 		return sdk.ZeroDec(), t.Exponent, types.ErrBlacklisted
 	}
 
-	// if a token is exempt from historic pricing, all price modes return Spot price.
-	// price mode last is unaffected.
-	if t.HistoricMedians == 0 && mode != types.PriceModeLast {
-		mode = types.PriceModeSpot
+	// if a token is exempt from historic pricing, all price modes ignore historic prices
+	// and use spot prices instead, sometimes also allowing expired prices.
+	if t.HistoricMedians == 0 {
+		mode = mode.IgnoreHistoric()
 	}
 
 	var price, historicPrice sdk.Dec
@@ -43,8 +43,8 @@ func (k Keeper) TokenPrice(ctx sdk.Context, baseDenom string, mode types.PriceMo
 		if err != nil {
 			return sdk.ZeroDec(), t.Exponent, errors.Wrap(err, "oracle")
 		}
-		if mode != types.PriceModeLast {
-			// unless price mode is last price, require spot price to be recent
+		if !mode.AllowsExpired() {
+			// with the exception of account summary queries, require spot prices to be recent
 			moduleTime := k.getLastInterestTime(ctx)
 			priceTime := spotPrice.Timestamp.Unix()
 			priceAge := moduleTime - priceTime
@@ -55,8 +55,8 @@ func (k Keeper) TokenPrice(ctx sdk.Context, baseDenom string, mode types.PriceMo
 		}
 	}
 
-	if mode != types.PriceModeSpot && mode != types.PriceModeLast {
-		// historic price is required for modes other than spot and last
+	if mode != types.PriceModeSpot && mode != types.PriceModeQuery {
+		// historic price is required for modes other than spot and query
 		var numStamps uint32
 		historicPrice, numStamps, err = k.oracleKeeper.MedianOfHistoricMedians(
 			ctx, strings.ToUpper(t.SymbolDenom), uint64(t.HistoricMedians))
@@ -73,13 +73,13 @@ func (k Keeper) TokenPrice(ctx sdk.Context, baseDenom string, mode types.PriceMo
 	}
 
 	switch mode {
-	case types.PriceModeSpot, types.PriceModeLast:
+	case types.PriceModeSpot, types.PriceModeQuery:
 		price = spotPrice.Rate
 	case types.PriceModeHistoric:
 		price = historicPrice
-	case types.PriceModeHigh:
+	case types.PriceModeHigh, types.PriceModeQueryHigh:
 		price = sdk.MaxDec(spotPrice.Rate, historicPrice)
-	case types.PriceModeLow:
+	case types.PriceModeLow, types.PriceModeQueryLow:
 		price = sdk.MinDec(spotPrice.Rate, historicPrice)
 	default:
 		return sdk.ZeroDec(), t.Exponent, types.ErrInvalidPriceMode.Wrapf("%d", mode)