+++ CHANGELOG.md
@@ -46,11 +46,14 @@ Ref: https://keepachangelog.com/en/1.0.0/
 
 ## [Unreleased]
 
+### API-Breaking
+- [2267](https://github.com/umee-network/umee/pull/2267) `BorrowLimit` field in QueryAccountSummaryResponse can be nil on missing borrow price (behavior now matches `LiquidationThreshold` field)
+
 ### Improvements
 
 - [2261](https://github.com/umee-network/umee/pull/2261) Use go 1.21
 - [2267](https://github.com/umee-network/umee/pull/2267) Leverage transactions accept spot prices up to 3 minutes old, and leverage queries use most recent spot price when required.
-- [2263](https://github.com/umee-network/umee/pull/2263) Add last price fields to account summary, and ensure all other fields use leverage logic prices.
+- [2263](https://github.com/umee-network/umee/pull/2263) Add spot price fields to account summary.
 
 ### Bug Fixes